improve PropertyImage image loading/rerendering
- Marginally optimize perfz with useCallback
- Rearrange <image> props so changing imageUrl causes rerender

This second bit is the important part: before, the component only rerendered when `url` changed (as
in the state variable), with nothing to listen for when the `imageUrl` value it gets initialized
with changed. The result was this shitty transitional edge case, wherein the placeholder would be
swapped out as soon as the component rerendered, but before the actual image url was ready, so you
invariably got a broken image until the first time you hover, replacing that invalid transitional
state with the onHover toggle business. Which is of course completely obtuse UX.